Замените все M_EVENT на прямые вызовы. Предфинал удаления xrMesssage

+++ trunk/xrGame/Actor.h
@@ -660,7 +660,7 @@ protected:
 		Fvector							m_AutoPickUp_AABB;
 		Fvector							m_AutoPickUp_AABB_Offset;
 public:
-		void							SetWeaponHideState				(u32 State, bool bSet);
+	void SetWeaponHideState(u16 State, bool bSet);
 		virtual CCustomOutfit*			GetOutfit() const;
 private:
 	CActorCondition				*m_entity_condition;
